install icons in application datadir
$prefix/share/icons is generally used by XDG icon themes, and the icons
installed there by convertall are not structured as such. Since they are
application private data rather than public icons, then install them
directly inside the data directory of convertall.

This removes the way to setup a custom path for them: it is not needed
anymore, as the loading code already looks for an 'icons' subdirectory
of the directory where the Python modules are.